Fixing bug concerning manual quality gates.
Approving a manual gate resulted in automated approvement of the same
gate in the next built. This was fixed by swapping to not approved after
one successful run in a manual gate. Manual checks can now be seen as
flip flop checks.

Execution checks did not understand build variables. This was fixed by
delegating command execution to a shell object which is informed about
build variables. These checks now write to the default build log.

Saving configuration to a custom xml file was also implemented.