Map build ABORTED to FAILURE of corresponding check result
In MavenBuildSuccessCheck the result of the check always was equal to the result of the build. Nevertheless the GUI components did not correctly visualize checks and gates which have been aborted. Due to this reason, the ABORTED result, which currently only occurs in MavenBuildSuccessCheck, is now remapped to FAILURE of the corresponding check.