Remove unused variable
To check whether a XPath expression in XMLCheck is syntactically valid, it has to be compiled and occurring errors have to be caught, as no API exists which allows testing syntactic validity directly. Nevertheless it is not necessary to save the result of the compilation in a variable which is not used afterwards.